requirements.txt
alembic==0.5.0
blessings==1.5
Chameleon==2.11
decorator==3.4.0
dogpile.cache==0.4.3
dogpile.core==0.4.1
Mako==0.8.0
MarkupSafe==0.15
parse==1.6.1
PasteDeploy==1.5.0
psycopg2==2.5
PyBrowserID==0.9.1
Pygments==1.6
pyramid-debugtoolbar==1.0.6
pyramid-persona==1.3.1
pyramid-tm==0.7
pyramid==1.4
repoze.lru==0.6
requests==1.2.0
six==1.3.0
SQLAlchemy==0.8.0
transaction==1.4.1
translationstring==1.1
venusian==1.0a8
waitress==0.8.2
WebOb==1.2.3
zope.deprecation==4.0.2
zope.interface==4.0.5
zope.sqlalchemy==0.7.2
